Tweaks to test-java-files.sh

+++ test-java-files.sh
@@ -1,5 +1,7 @@
 #!/usr/bin/env bash
 
+set -euo pipefail
+
 JAR=~/hacks/bhs-cs/java/target/bhs-cs.jar
 RUNNER=com.gigamonkeys.bhs.testing.TestRunner
 
@@ -8,10 +10,15 @@ shift
 
 for f; do
     echo -n "Testing $f ... "
-    if java -cp "$JAR" "$RUNNER" "$tester" "$f" > "$(dirname "$f")/results.json" 2> /dev/null; then
-        echo "ok."
+    rm -f "$(dirname "$f")/broken.txt"
+    if [[ ! -e "$(dirname "$f")/missing.txt" ]]; then
+        if java -cp "$JAR" "$RUNNER" "$tester" "$f" > "$(dirname "$f")/results.json" 2> /dev/null; then
+            echo "ok."
+        else
+            echo "broken."
+            touch "$(dirname "$f")/broken.txt"
+        fi
     else
-        echo "broken."
-        touch "$(dirname "$f")/broken.txt"
+        echo "missing."
     fi
 done
